Refactored the data processors and added some tests
